requirements.txt
python-dotenv>=1.0.0
groq>=0.28.0
langchain-groq>=0.2.0
langchain-core>=0.3.40
langgraph>=0.4.9
langgraph-prebuilt>=0.2.2
langgraph-checkpoint>=2.1.0
langgraph-supervisor>=0.0.27
pydantic<2.10,>=2.9

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
